--- a/tools/train.py
+++ b/tools/train.py
@@ -94,34 +94,32 @@ def parse_args():
 def setup_multi_processes(cfg):
     # set multi-process start method as `fork` to speed up the training
     if platform.system() != 'Windows':
-        mp.set_start_method('fork')
+        mp_start_method = cfg.get('mp_start_method', 'fork')
+        mp.set_start_method(mp_start_method)
 
-    # disable opencv multithreading to avoid overwheleming
-    cv2.setNumThreads(0)
+    # disable opencv multithreading to avoid system being overloaded
+    opencv_num_threads = cfg.get('opencv_num_threads', 0)
+    cv2.setNumThreads(opencv_num_threads)
 
     # setup OMP threads
     # This code is referred from https://github.com/pytorch/pytorch/blob/master/torch/distributed/run.py  # noqa
     if ('OMP_NUM_THREADS' not in os.environ and cfg.data.workers_per_gpu > 1):
         omp_num_threads = 1
         warnings.warn(
-            f'\n*****************************************\n'
             f'Setting OMP_NUM_THREADS environment variable for each process '
             f'to be {omp_num_threads} in default, to avoid your system being '
             f'overloaded, please further tune the variable for optimal '
-            f'performance in your application as needed. \n'
-            f'*****************************************')
+            f'performance in your application as needed.')
         os.environ['OMP_NUM_THREADS'] = str(omp_num_threads)
 
     # setup MKL threads
     if 'MKL_NUM_THREADS' not in os.environ and cfg.data.workers_per_gpu > 1:
         mkl_num_threads = 1
         warnings.warn(
-            f'\n*****************************************\n'
             f'Setting MKL_NUM_THREADS environment variable for each process '
             f'to be {mkl_num_threads} in default, to avoid your system being '
             f'overloaded, please further tune the variable for optimal '
-            f'performance in your application as needed. \n'
-            f'*****************************************')
+            f'performance in your application as needed.')
         os.environ['MKL_NUM_THREADS'] = str(mkl_num_threads)