@shinya7y 2022-01-14
Does `setup_multi_processes` also work for `tools/test.py`?