@RangiLyu 2022-01-17
> Does `setup_multi_processes` also work for `tools/test.py`?

Yes, it works. I'll add this both in train.py and test.py.